requirements.txt
flask
flask_sqlalchemy
flask_script
sqlalchemy_utils
python-dotenv
gunicorn